+++ src/Components/Display/ProgressBar/index.ts
@@ -1,0 +1,3 @@
+import UpProgressBar from './UpProgressBar';
+
+export default UpProgressBar;